major changes: login/register employee/admin, different page views per role, edit button disabled for employee, admin can delete orders and register new employee, some changes in project structure, xml builder done (todo: edit/update saved units)

+++ src/main/resources/static/js/orders-list.js
@@ -1,9 +1,36 @@
 $(function() {
 
     $(document).ready(function() {
-        loadOrders("/all")
-        localStorage.clear();
+        if (checkIfLoggedIn()) {
+            $('#login-li').attr("hidden", true);
+            $('#logout-li').removeAttr("hidden");
+            if (sessionStorage.getItem("role") === 'ADMIN') {
+                $("#register-li").removeAttr("hidden");
+            }
+            loadOrders("/all");
+            localStorage.clear();
+        } else {
+            alert("Your session is either expired or you're not logged in");
+            window.location.href = "/login.html";
+        }
     });
+
+    setInterval(function() {
+        $.get('/check', {"token" : sessionStorage.getItem("token")}, function(res) {
+            if (res === false) {
+                $('#register-li').attr("hidden", true);
+                $('#login-li').removeAttr("hidden");
+                $('#logout-li').attr("hidden", true);
+                sessionStorage.clear();
+                alert("Your session is either expired or you're not logged in");
+                window.location.href = "/login.html";
+            }
+        });
+    }, 1000 * 60);
+
+    let checkIfLoggedIn = function() {
+        return sessionStorage.getItem("token") !== null;
+    }
 
     let getResponseElement = function(response) {
         let tbody = $(".orders");
@@ -17,37 +44,89 @@ $(function() {
             response.customerName + "</td>"));
         trow.append($("<td class=\"px-6 py-4\" id=\"status-order-" + response.id + "\">" +
             response.status + "</td>"));
-        trow.append($("<td class=\"px-6 py-4\">" +
-            "<button id='edit-id-" + response.id + "' type='submit' class='edit-button'>Edit</button></td>"));
+        checkRoles(trow, response);
         return trow;
     };
 
+    let checkRoles = function(trow, response) {
+        if (response.status === 'FINISHED' && sessionStorage.getItem("role") === 'EMPLOYEE') {
+            trow.append($("<td class=\"px-6 py-4\">" +
+                "<button id='edit-id-" + response.id + "' disabled='disabled' style='color : gray' type='submit' class='edit-button'>Edit</button></td>"));
+        } else {
+            trow.append($("<td class=\"px-6 py-4\">" +
+                "<button id='edit-id-" + response.id + "' type='submit' class='edit-button'>Edit</button></td>"));
+        }
+        if (sessionStorage.getItem("role") === 'ADMIN') {
+            trow.append($("<td class=\"px-6 py-4\">" +
+                "<button id='edit-id-" + response.id + "' type='submit' style='color: red' class='delete-button'>Delete</button></td>"));
+        } else {
+            trow.append($("<td class=\"px-6 py-4\">" +
+                "<button disabled='disabled' type='submit' style='color: gray'>Delete</button></td>"));
+        }
+    }
+
     let loadOrders = function(url) {
-        $.get(url, {}, function(response) {
-            let tbody = $(".orders");
-            tbody.empty();
-            for (i in response) {
-                let res = getResponseElement(response[i]);
-                tbody.append(res);
-            }
-        });
+        if (checkIfLoggedIn()) {
+            $.get(url, {}, function(response) {
+                let tbody = $(".orders");
+                tbody.empty();
+                for (i in response) {
+                    let res = getResponseElement(response[i]);
+                    tbody.append(res);
+                }
+            });
+        } else {
+            alert("Your session is either expired or you're not logged in");
+        }
     }
 
     let loadOrderBySearch = function(id) {
-        $.get("/search/" + id, {}, function(response) {
-            let tbody = $(".orders");
-            tbody.empty();
-            let res = getResponseElement(response);
-            tbody.append(res);
-        });
+        if (checkIfLoggedIn()) {
+            $.get("/search/" + id, {}, function(response) {
+                let tbody = $(".orders");
+                tbody.empty();
+                let res = getResponseElement(response);
+                tbody.append(res);
+            });
+        } else {
+            alert("Your session is either expired or you're not logged in");
+        }
     }
 
-    $('tbody').on('click', '.edit-button',  function() {
+    $('#register-p').on('click', function() {
+        window.location.href = "/register.html";
+    });
+
+    $('#logout-p').on('click', function() {
+        sessionStorage.clear();
+        window.location.href = "/login.html";
+    });
+
+    $('tbody')
+        .on('click', '.edit-button',  function() {
         let stringId = this.id;
         let id = stringId.substring(stringId.length - 1);
         localStorage.setItem("id", id);
         console.log(id);
-        window.location.replace("/orders-form.html");
+        window.location.href = "/orders-form.html";
+    })
+        .on('click', '.delete-button',  function() {
+        let stringId = this.id;
+        let id = stringId.substring(stringId.length - 1);
+        let result = confirm("Are you sure you want to delete order #" + id);
+        if (result) {
+            $.ajax({
+                "url": '/delete/' + id,
+                "type": 'DELETE',
+                "success": function(result) {
+                    alert("Order #" + result + " was successfully deleted!");
+                    location.reload();
+                },
+                "error": function () {
+                    alert("Error deleting order #" + id);
+                }
+            });
+        }
     });
 
     $('#all').on('click', function() {
@@ -65,5 +144,9 @@ $(function() {
     $('#search-button').on('click', function() {
         let input = $('#input-placeholder').val();
         loadOrderBySearch(input);
-    })
+    });
+
+    $('#login-p').on('click', function() {
+        window.location.href = "/login.html";
+    });
 })
